default to a common curve if no curves advertised
The ECC TLS RFC states that if the client doesn't advertise
any curves, it means that it supports any and the server
can then select any curve. That was OK when there was only
one set of curves supported in TLS. Now we have the legacy
curves, brainpool curves and djb (x25519 and x448) curves,
meaning that client that doesn't advertise any curves is likely
to support only the legacy curves, not the newer ones.

Thus we need to separate the default curve from most wanted
curve as the former must be most compatible while still
relatively secure while the latter can be fastest or most
secure, but will be negotiated only when client also supports
it.